+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 - Change SteamOS packages to be optionally installed as part of the first-time setup.
 - Add support for the Vulkan graphics pipeline library on Intel.
 - Add support for older integrated Intel graphics.
+- Add FileZilla as an FTP client.
 
 ## 3.2.1 - The SteamOS 3.4 Update